LayoutTests/imported/w3c/web-platform-tests/css/css-shadow-parts/invalidation-change-part-name-idl-domtokenlist-expected.txt

The following text should be green:

FAIL Part in selected host changed color via part IDL DOMTokenList attribute. assert_equals: expected "rgb(0, 128, 0)" but got "rgb(0, 0, 255)"
